+++ app/src/serializers/area.serializer.js
@@ -10,6 +10,7 @@ const areaSerializer = new JSONAPISerializer('area', {
         'createdAt',
         'updatedAt',
         'image',
+        'env',
         'datasets',
         'use',
         'iso',